split "dice admin *" to "dicectl *"
