+++ src/dice_cli/main.py
@@ -5,8 +5,8 @@ import dice_lib.parameters as dice_params
 import rich
 import typer
 
-from . import __version__, _date, admin, benchmark, check, docs, info, job
-from .logger import admin_logger, console_handler, user_logger
+from . import __version__, _date, benchmark, check, docs, info, job
+from .logger import console_handler, user_logger
 
 
 def user_callback(
@@ -23,22 +23,7 @@ def user_callback(
         console_handler.setLevel(logging.DEBUG)
 
 
-def admin_callback(
-    verbose: bool = typer.Option(False, "--verbose", "-v", help="Verbose output"),
-    debug: bool = typer.Option(False, "--debug", "-d", help="Debug output"),
-) -> Any:
-    # verbose is debug
-    verbose = debug or verbose
-    debug = verbose
-
-    if debug:
-        admin_logger.setLevel(logging.DEBUG)
-        # workaround for rich
-        console_handler.setLevel(logging.DEBUG)
-
-
 app = typer.Typer()
-app.add_typer(admin.app, name="admin", callback=admin_callback)
 app.add_typer(benchmark.app, name="benchmark", callback=user_callback)
 app.add_typer(check.app, name="check", callback=user_callback)
 app.add_typer(docs.app, name="docs", callback=user_callback)
